ext-image-capture-source-v1: Send all render formats
ext-image-copy-capture clients receive a list of compatible formats for
a capture session of a given source, which they will use to allocate a
buffer that we will copy our output buffer to on every captured frame.

We were only reporting the current swapchain format, which is useful if
we want to optimize for a direct copy instead of a render, but comes at
the cost of incompatibility if the capture sink does not support that
specific format. Chromium's libwebrtc (also used by Firefox) does for
exmaple not support our 10-bit format choices, breaking 10-bit screen
capture.

Instead, advertise all formats supported by our renderer and let the
capture sink pick its favorite. We retain the current behavior for
shm to avoid paying conversion costs if using pixman renderer.